Writing setting-win and code-generation
